+++ app/Http/Controllers/SubjectController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\SubjectRequest;
+use App\Models\Student;
 use App\Models\Subject;
 
 class SubjectController extends Controller
@@ -13,6 +14,11 @@ class SubjectController extends Controller
     public function index()
     {
         $subjects = Subject::paginate(5);
+
+        if (request()->wantsJson()) {
+            return response()->json($subjects);
+        }
+
         return inertia('Subjects/Index', ['subjects' => $subjects]);
     }
 
@@ -68,7 +74,23 @@ class SubjectController extends Controller
      */
     public function destroy(Subject $subject)
     {
+        $hasStudents = Student::whereHas('subjects', function ($query) use ($subject) {
+            $query->where('subject_id', $subject->id);
+        })->exists();
+
+        if ($hasStudents) {
+            return response()->json(['error' => 'This subject has associated students and cannot be eliminated.'], 422);
+        }
+
         $subject->delete();
-        return redirect()->route('subject.index');
+        return response()->json(['message' => 'Subject successfully deleted.']);
+    }
+
+    public function getSubjectsWithProfessors()
+    {
+        // Get only subjects with an assigned professor
+        $subjectsWithProfessors = Subject::has('professors')->get();
+
+        return response()->json($subjectsWithProfessors);
     }
 }
